review fixes and moving dialect stats to new struct
