+++ src/config.h
@@ -133,8 +133,6 @@ typedef struct {
   // Can allow to control the seperation between phrases in different array slots (related to the SLOP parameter in ft.search command)
   // Default value is 100. 0 will not increment (as if all text is a continus phrase).
   unsigned int multiTextOffsetDelta;
-  // bitarray of dialects used by all indices
-  uint_least8_t used_dialects;
   // The number of iterations to run while performing background indexing
   // before we call usleep(1) (sleep for 1 micro-second) and make sure that
   // we allow redis process other commands.
@@ -213,8 +211,6 @@ sds RSConfig_GetInfoString(const RSConfig *config);
 
 void RSConfig_AddToInfo(RedisModuleInfoCtx *ctx);
 
-void DialectsGlobalStats_AddToInfo(RedisModuleInfoCtx *ctx);
-
 #ifdef MT_BUILD
 void UpgradeDeprecatedMTConfigs();
 #endif
@@ -226,11 +222,6 @@ void UpgradeDeprecatedMTConfigs();
 #define DEFAULT_FORK_GC_RUN_INTERVAL 30
 #define SEARCH_REQUEST_RESULTS_MAX 1000000
 #define NR_MAX_DEPTH_BALANCE 2
-#define MIN_DIALECT_VERSION 1 // MIN_DIALECT_VERSION is expected to change over time as dialects become deprecated.
-#define MAX_DIALECT_VERSION 4 // MAX_DIALECT_VERSION may not exceed MIN_DIALECT_VERSION + 7.
-#define DIALECT_OFFSET(d) (1ULL << (d - MIN_DIALECT_VERSION))// offset of the d'th bit. begins at MIN_DIALECT_VERSION (bit 0) up to MAX_DIALECT_VERSION.
-#define GET_DIALECT(barr, d) (!!(barr & DIALECT_OFFSET(d)))  // return the truth value of the d'th dialect in the dialect bitarray.
-#define SET_DIALECT(barr, d) (barr |= DIALECT_OFFSET(d))     // set the d'th dialect in the dialect bitarray to true.
 #define VECSIM_DEFAULT_BLOCK_SIZE   1024
 #define DEFAULT_MIN_STEM_LENGTH 4
 #define MIN_MIN_STEM_LENGHT 2 // Minimum value for minStemLength
@@ -280,7 +271,6 @@ void UpgradeDeprecatedMTConfigs();
     .requestConfigParams.dialectVersion = 1,                                                                          \
     .vssMaxResize = 0,                                                                                                \
     .multiTextOffsetDelta = 100,                                                                                      \
-    .used_dialects = 0,                                                                                               \
     .numBGIndexingIterationsBeforeSleep = 100,                                                                        \
     .prioritizeIntersectUnionChildren = false                                                                         \
   }
