+++ src/global_stats.h
@@ -8,6 +8,10 @@
 
 #include "spec.h"
 
+#define DIALECT_OFFSET(d) (1ULL << (d - MIN_DIALECT_VERSION))// offset of the d'th bit. begins at MIN_DIALECT_VERSION (bit 0) up to MAX_DIALECT_VERSION.
+#define GET_DIALECT(barr, d) (!!(barr & DIALECT_OFFSET(d)))  // return the truth value of the d'th dialect in the dialect bitarray.
+#define SET_DIALECT(barr, d) (barr |= DIALECT_OFFSET(d))     // set the d'th dialect in the dialect bitarray to true.
+
 typedef struct {
   size_t numTextFields;
   size_t numTextFieldsSortable;
@@ -31,8 +35,9 @@ typedef struct {
 } FieldsGlobalStats;
 
 typedef struct {
-  size_t total_unique_queries;  // Number of unique queries, not counting `FT.CURSOR READ` (an iteration of a previous query)
-  size_t total_query_commands;  // Number of total query commands, including `FT.CURSOR READ`
+  size_t total_queries_processed; // Number of successful queries. If using cursors, not counting reading from the cursor
+  size_t total_query_commands;    // Number of successful query commands, including `FT.CURSOR READ`
+  uint_least8_t used_dialects;    // bitarray of dialects used by all indices
 } TotalGlobalStats;
 
 // The global stats object type
@@ -58,3 +63,11 @@ void FieldsGlobalStats_AddToInfo(RedisModuleInfoCtx *ctx);
  * Increase all relevant counters in the global stats object.
  */
 void TotalGlobalStats_CountQuery(uint32_t reqflags);
+
+/**
+ * Add all the query-related information to the INFO command.
+ */
+void TotalGlobalStats_Queries_AddToInfo(RedisModuleInfoCtx *ctx);
+
+
+void DialectsGlobalStats_AddToInfo(RedisModuleInfoCtx *ctx);
