+++ src/info_command.c
@@ -13,6 +13,7 @@
 #include "geometry_index.h"
 #include "redismodule.h"
 #include "reply_macros.h"
+#include "global_stats.h"
 
 static void renderIndexOptions(RedisModule_Reply *reply, IndexSpec *sp) {
